Rewrite the computation of flat and island perimeters.
Use the new "contour" algorithm for both. This does not depend on the vertex
coordinates nor on the VertexIndex.

Previously, `flat_contour` used the `VertexIndex` to build the perimeter
order, but that would panic if the flat-extended face had mixed textures
or different normals.

It worked anyways because in an OBJ file an n-gon can only have one texture
and one set of normals, so this situation was hard to build.